refactor: useAnimateContainer 훅 분리
